feat: add game category validation and room owner management
