--- src/main/java/com/icebreaker/be/global/annotation/CurrentGameCategory.java
@@ -1,13 +1,0 @@
-package com.icebreaker.be.global.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface CurrentGameCategory {
-
-    String roomCodeParam() default "roomCode";
-}
